added authentication

+++ src/components/WorkoutDetails.jsx
@@ -1,21 +1,27 @@
 import React from "react";
 import { useQueryClient } from "react-query";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 const WorkoutDetails = ({ workout }) => {
   //! REACT QUERY
   const queryClient = useQueryClient();
+  const { user } = useAuthContext();
   const handleClick = async (e) => {
     e.preventDefault();
-
+    console.log(user.token);
+    if (!user) {
+      return;
+    }
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: { Authorization: `Bearer ${user.token}` },
     });
 
     //! REACT QUERY
 
     if (!response.ok) {
-      console.log("Error");
+      console.log("Error", response);
     }
     if (response.ok) {
       queryClient.invalidateQueries("workoutData");
